Fix remaining build errors in AI Insights Phase 11/12
- Fix AIInsightsConfigRootView.swift: Change MarkdownView to RichMarkdownView
- Fix HealthDataExporter.swift: Handle whyHighLow case properly in switch
- Move WhyHighLowBannerView into HomeRootView.swift (was in separate file not in Xcode project)
- Delete redundant WhyHighLowBannerView.swift file
